+++ simple_manga_downloader/modules/manganelo_com.py
@@ -26,10 +26,10 @@ class Manganelo:
 
     @request_exception_handler
     def get_main(self, title_return=False):
-        '''
+        """
         Gets the main manga info like title, cover url and chapter links
         title_return=True will only get the title and return
-        '''
+        """
         r = self.session.get(self.manga_link, timeout=5)
         r.raise_for_status()
         soup = BeautifulSoup(r.text, "html.parser")
@@ -44,9 +44,9 @@ class Manganelo:
         return True
 
     def get_chapters(self):
-        '''
+        """
         Handles the chapter data by assigning chapter numbers
-        '''
+        """
         for chapter in self.data:
 
             search_reg = re.search(r" (\d+\.?\d*)(?:: (.*))?", chapter.text)
@@ -67,9 +67,9 @@ class Manganelo:
 
     @request_exception_handler
     def get_info(self, ch):
-        '''
+        """
         Gets the needed data abut the chapters from the site
-        '''
+        """
         link = self.chapters[ch]["link"]
 
         r = self.session.get(link, timeout=5)