+++ simple_manga_downloader/modules/mangatown_com.py
@@ -26,10 +26,10 @@ class Mangatown():
 
     @request_exception_handler
     def get_main(self, title_return=False):
-        '''
+        """
         Gets the main manga info like title, cover url and chapter links
         title_return=True will only get the title and return
-        '''
+        """
         r = self.session.get(self.manga_link, timeout=5)
         r.raise_for_status()
         soup = BeautifulSoup(r.text, "html.parser")
@@ -45,9 +45,9 @@ class Mangatown():
         return True
 
     def get_chapters(self):
-        '''
+        """
         Gets the list of available chapters
-        '''
+        """
         for chapter in self.data:
             chapter_link = f"{self.base_link}{chapter.find('a')['href']}"
             try:
@@ -74,7 +74,7 @@ class Mangatown():
 
     @request_exception_handler
     def get_info(self, ch):
-        '''Gets the needed data abut the chapters from the site'''
+        """Gets the needed data abut the chapters from the site"""
 
         r = self.session.get(self.chapters[ch]["link"], timeout=5)
         r.raise_for_status()